requirements.txt
discord~=2.0.0
numpy~=1.23.4
pandas~=1.5.1
python-dotenv~=0.21.0